refactor: 자동차 시도 횟수만큼 전진 횟수 카운트
